Chore: 리팩토링 완료
