+++ lib/utils/logger_utils.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/foundation.dart';
+
+void safePrint(Object? object) {
+  if (!kReleaseMode) {
+    print(object);
+  }
+}
